requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==25.1.0
packaging==26.0
sqlparse==0.5.5
tzdata==2025.3
dj-database-url==2.1.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
